Make nothing :: Maybe a, remove extra parameter
`nothing` is only used in one place, and as a function it was called
incorrectly before: `nothing()` instead of `nothing(undefined)`.

This commit fixes that by removing the extra layer of function wrapping
entirely, and making `nothing` a constant of type `Maybe a`. This is to
avoid the more verbose alternative - adding the `undefined` argument.